+++ stepik-java-api/src/main/java/org/stepik/api/objects/enrollments/EnrollmentPost.java
@@ -4,13 +4,13 @@ package org.stepik.api.objects.enrollments;
  * @author meanmail
  */
 public class EnrollmentPost {
-    private int course;
+    private long course;
 
-    public int getCourse() {
+    public long getCourse() {
         return course;
     }
 
-    public void setCourse(int course) {
+    public void setCourse(long course) {
         this.course = course;
     }
 }
